Move glinstance to pkg
