+++ commands/api/api.go
@@ -22,8 +22,8 @@ import (
 	"github.com/MakeNowJust/heredoc"
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/glinstance"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/glinstance"
 	"github.com/spf13/cobra"
 	jsonPretty "github.com/tidwall/pretty"
 	"github.com/xanzy/go-gitlab"
