+++ commands/cmdutils/remote_resolver.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/glinstance"
 	"github.com/profclems/glab/internal/glrepo"
 	"github.com/profclems/glab/pkg/git"
+	"github.com/profclems/glab/pkg/glinstance"
 )
 
 type remoteResolver struct {
