+++ commands/release/view/view.go
@@ -8,8 +8,8 @@ import (
 	"github.com/profclems/glab/commands/cmdutils"
 	"github.com/profclems/glab/commands/release/releaseutils"
 	"github.com/profclems/glab/internal/config"
-	"github.com/profclems/glab/internal/glinstance"
 	"github.com/profclems/glab/internal/glrepo"
+	"github.com/profclems/glab/pkg/glinstance"
 	"github.com/profclems/glab/pkg/iostreams"
 	"github.com/profclems/glab/pkg/utils"
 	"github.com/spf13/cobra"
