+++ pkg/glinstance/host_test.go
